pdf_cropper/requirements.txt
pdf2image==1.17.0
Pillow==11.3.0
PyPDF2==3.0.1
